take out javascripts
