+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "build": "tsc -p tsconfig.json",
-    "start": "node index.js"
+    "start": "node index.js",
+    "dev": "ts-node ./src/index.ts"
   },
   "keywords": [],
   "author": "",
@@ -22,6 +23,7 @@
   "devDependencies": {
     "@types/node": "^16.11.11",
     "sequelize-cli": "^6.3.0",
+    "ts-node": "^10.4.0",
     "typescript": "^4.5.2"
   }
 }
