--- src/db/config/connection.js
@@ -1,15 +1,0 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-const config_1 = __importDefault(require("./config"));
-const initializeDatabase = async () => {
-    try {
-        await config_1.default.sync();
-    }
-    catch (err) {
-        console.error(err);
-    }
-};
-exports.default = initializeDatabase;
